refactor: 🧹 remove unused SUPPORTED_FORMATS import in frame_extractor.py
